test: add tests for `Options`
This extracts the class into its own file as well.
