+++ src/index.ts
@@ -1,10 +1,8 @@
-import { existsSync, readFileSync, writeFileSync } from 'fs';
-import { basename, extname, resolve } from 'path';
-import { sync as resolveSync } from 'resolve';
-import iterateSources, { PathPredicate } from './iterateSources';
-import TransformRunner, { Plugin, Source, SourceTransformResult } from './TransformRunner';
-
-const DEFAULT_EXTENSIONS = new Set(['.js', '.jsx']);
+import { writeFileSync } from 'fs';
+import { basename } from 'path';
+import iterateSources from './iterateSources';
+import Options, { DEFAULT_EXTENSIONS } from './Options';
+import TransformRunner, { Source, SourceTransformResult } from './TransformRunner';
 
 function printHelp(out: NodeJS.WritableStream) {
   let $0 = basename(process.argv[1]);
@@ -105,155 +103,6 @@ export default async function run(args: Array<string>) {
   }
 
   process.exit(hasErrors ? 1 : 0);
-}
-
-type ParseOptionsResult = Options | Error;
-
-class Options {
-  private plugins?: Array<Plugin>;
-
-  constructor(
-    readonly sourcePaths: Array<string>,
-    readonly pluginFilePaths: Array<string>,
-    readonly pluginOptions: Map<string, object>,
-    readonly extensions: Set<string>,
-    readonly requires: Array<string>,
-    readonly ignore: PathPredicate,
-    readonly stdio: boolean,
-    readonly help: boolean,
-  ) {}
-
-  getPlugins(): Array<Plugin> {
-    if (!this.plugins) {
-      this.plugins = this.loadPlugins();
-    }
-
-    return this.plugins;
-  }
-
-  loadRequires() {
-    for (let modulePath of this.requires) {
-      require(modulePath);
-    }
-  }
-
-  private loadPlugins(): Array<Plugin> {
-    return this.pluginFilePaths.map(pluginFilePath => {
-      let name = basename(pluginFilePath, extname(pluginFilePath));
-      let options = this.pluginOptions.get(name);
-      let exports = require(pluginFilePath);
-      let plugin;
-
-      if (exports.default) {
-        plugin = exports.default;
-      } else {
-        plugin = exports;
-      }
-
-      if (options) {
-        return [plugin, options];
-      } else {
-        return plugin;
-      }
-    });
-  }
-
-  static parse(args: Array<string>): ParseOptionsResult {
-    let sourcePaths: Array<string> = [];
-    let pluginFilePaths: Array<string> = [];
-    let pluginOptions: Map<string, object> = new Map();
-    let extensions = DEFAULT_EXTENSIONS;
-    let ignore = (path: string, basename: string, root: string) => basename[0] === '.';
-    let requires: Array<string> = [];
-    let stdio = false;
-    let help = false;
-
-    for (let i = 0; i < args.length; i++) {
-      let arg = args[i];
-
-      switch (arg) {
-        case '-p':
-        case '--plugin':
-          i++;
-          pluginFilePaths.push(getRequirableModulePath(args[i]));
-          break;
-
-        case '-o':
-        case '--plugin-options':
-          i++;
-          let nameAndOptions = args[i].split('=');
-          let name = nameAndOptions[0];
-          let optionsRaw = nameAndOptions[1];
-
-          if (optionsRaw && optionsRaw[0] === '@') {
-            optionsRaw = readFileSync(args[i].slice(1), { encoding: 'utf8' });
-            pluginOptions.set(name, JSON.parse(optionsRaw));
-          }
-
-          try {
-            pluginOptions.set(name, JSON.parse(optionsRaw));
-          } catch (err) {
-            throw new Error(`unable to parse JSON config for ${name}: ${optionsRaw}`);
-          }
-          break;
-
-        case '-r':
-        case '--require':
-          i++;
-          requires.push(getRequirableModulePath(args[i]));
-          break;
-
-        case '--extensions':
-          i++;
-          extensions = new Set(
-            args[i]
-              .split(',')
-              .map(ext => ext[0] === '.' ? ext : `.${ext}`)
-          );
-          break;
-
-        case '-s':
-        case '--stdio':
-          stdio = true;
-          break;
-
-        case '-h':
-        case '--help':
-          help = true;
-          break;
-
-        default:
-          if (arg[0] === '-') {
-            throw new Error(`unexpected option: ${arg}`);
-          } else {
-            sourcePaths.push(arg);
-          }
-          break;
-      }
-    }
-
-    return new Options(
-      sourcePaths,
-      pluginFilePaths,
-      pluginOptions,
-      extensions,
-      requires,
-      ignore,
-      stdio,
-      help
-    );
-  }
-}
-
-/**
- * Gets a path that can be passed to `require` for a given module path.
- */
-function getRequirableModulePath(modulePath: string): string {
-  if (existsSync(modulePath) || existsSync(modulePath + '.js')) {
-    return resolve(modulePath);
-  } else {
-    return resolveSync(modulePath, { basedir: process.cwd() });
-  }
 }
 
 /**
